refactor(automations): co-locate the Python-metric reserved-variable helper
`reservedPythonMetricVariablesForScope` is meaningful only to the
AddEditRuleDialog family, so `lib/` was the wrong home: exporting it there
advertises a rule-dialog contract as a repo-wide LLM utility. Move it and its
tests next to their only consumer, per the co-location rule in
.agents/skills/opik-frontend/code-quality.md ("move it to lib/ only when a
third unrelated consumer appears").

`resolveTraceEvaluatorVariableDefault` stays in lib/llm.ts: it predates this
PR on main and already serves both dialog components.
